requirements.txt
pandas
numpy
scikit-learn
spacy
sentence-transformers
umap-learn
hdbscan
plotly
